+++ src/Pages/Login.js
@@ -2,7 +2,6 @@ import { useAuth } from "../Contexts";
 import { useState } from "react";
 import { Link } from "react-router-dom";
 
-
 export function Login() {
 
     const [username, setUsername] = useState();
